restapi/requirements.txt
djangorestframework
Django
Pygments
Markdown
coreapi
psycopg2-binary
dj-database-url
gunicorn
whitenoise
celery==5.3.6
amqp==5.1.1
redis==4.5.5
numpy==1.24.3
opencv-python==4.8.0.74
onnx==1.14.0
insightface==0.7.3
psutil==5.9.5
tk==0.1.0
customtkinter==5.2.0
tkinterdnd2==0.3.0
onnxruntime==1.15.0
tensorflow==2.13.0
opennsfw2==0.10.2
protobuf==4.23.4
tqdm==4.65.0
google-api-core==2.11.1
google-api-python-client==2.98.0
google-auth==2.22.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.60.0
pytube==15.0.0
django-cors-headers
djangorestframework-simplejwt
flake8
black
isort
mypy
boto3